feat(CSS): Add JetBrains Mono font import and set it as base font family
This commit adds an `@import` statement to import the 'JetBrains Mono' font from Google Fonts. It also sets the imported font as the base font family for the project by updating the `--ifm-font-family-base` variable in the `:root` selector. This change aims to improve the overall typography and readability of the website, especially for code snippets and content-centric sections.