feat(models): Add new input and scores
This commit adds a new input and its corresponding toxicity scores to the models. The input is "I think you're adorable. Do you want to go on a date after work? Maybe we could even have sex after?!" and the scores for each model are as follows:
- Toxicity: 48.5
- SevereToxicity: 2.31
- IdentityAttack: 2.09
- Insult: 8.69
- Profanity: 36.3
- Threat: 2.46
- SexuallyExplicit: 57.42
- Flirtation: 96.79
- Inflammatory: 27.08
- Obscene: 39.24
- Spam: 10.75

These additions enhance the capabilities of our models, allowing users to test their own inputs and obtain relevant toxicity scores.

Please note that there is no newline at the end of the file in this commit.