refactor(docs): Rename file paths for metrics and models
This commit renames the file paths for the metrics and models documentation files. The original file paths contained an ampersand symbol (&), which has been replaced with the word "and" to improve readability and adhere to naming conventions. The changes include renaming "_category_.json", "metrics.mdx", and "models.mdx" files in the "docs/metrics & models" directory to "docs/metrics and models".